feature(like-dislike): add more tests
